# resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "El login exitoso de AuthService devuelve un Usuario con un campo \"token\".\nDicho token se debe incluir como header de las llamadas del resto de las operaciones de cada servicio.",
    "version" : "1.0-SNAPSHOT",
    "title" : "Sulfur backend",
    "contact" : {
      "name" : "rsperoni@magnesium.coop"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/api",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/rubros" : {
      "get" : {
        "tags" : [ "rubros" ],
        "summary" : "Get rubros",
        "description" : "",
        "operationId" : "findAll",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Rubro"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "rubros" ],
        "summary" : "Create Rubro",
        "description" : "",
        "operationId" : "create",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Rubro"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rubro"
            }
          }
        }
      }
    },
    "/rubros/{id}" : {
      "get" : {
        "tags" : [ "rubros" ],
        "summary" : "Get Rubro",
        "description" : "",
        "operationId" : "find",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rubro"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "rubros" ],
        "summary" : "Edit Rubro",
        "description" : "",
        "operationId" : "edit",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Rubro"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rubro"
            }
          },
          "304" : {
            "description" : "Error: objeto no modificado"
          }
        }
      },
      "delete" : {
        "tags" : [ "rubros" ],
        "summary" : "Delete Rubro",
        "description" : "",
        "operationId" : "delete",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rubro"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "LocalDateTime" : {
      "type" : "object",
      "properties" : {
        "nano" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dayOfMonth" : {
          "type" : "integer",
          "format" : "int32"
        },
        "monthValue" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dayOfYear" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dayOfWeek" : {
          "type" : "string",
          "enum" : [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
        },
        "year" : {
          "type" : "integer",
          "format" : "int32"
        },
        "month" : {
          "type" : "string",
          "enum" : [ "JANUARY", "FEBRUARY", "MARCH", "APRIL", "MAY", "JUNE", "JULY", "AUGUST", "SEPTEMBER", "OCTOBER", "NOVEMBER", "DECEMBER" ]
        },
        "hour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "minute" : {
          "type" : "integer",
          "format" : "int32"
        },
        "second" : {
          "type" : "integer",
          "format" : "int32"
        },
        "chronology" : {
          "$ref" : "#/definitions/Chronology"
        }
      }
    },
...
    "Chronology" : {
      "type" : "object",
      "properties" : {
        "calendarType" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        }
      }
    },
...
    "Rubro" : {
      "type" : "object",
      "required" : [ "nombre" ],
      "properties" : {
        "creado" : {
          "$ref" : "#/definitions/LocalDateTime"
        },
        "modificado" : {
          "$ref" : "#/definitions/LocalDateTime"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nombre" : {
          "type" : "string"
        },
        "descripcion" : {
          "type" : "string"
        }
      }
    },
...
  }
}